Lots of changes this time
FLushed out the inspect_ingredients() dish method
to allow the user to navigate through some sub
menus to well, inspect some ingredients. Note that
you can't "go back" to the main menu after initially
selecting a type of ingredient-- this is by design.

randomize_contents was also tweaked as it was discovered that
if there were multiple ingreidents of the same type in the fridge,
they would all have the exact same freshness value. This has been fixed.

The game is now 100% functional, as initialy intended.